# wms_api/api-contracts.yaml
openapi: 3.0.3
info:
  title: WMS API - Warehouse Management System
  description: |
    API контракты для системы управления складом (WMS).
    Включает управление товарами, заказами, размещением, пользователями и логистикой доставки.
  version: 1.0.0
  contact:
    name: WMS Development Team
servers:
  - url: http://localhost:3002/api
    description: Development server (WMS API)
  - url: http://localhost:3001/api
    description: Mock server (wms_site/server)
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: Получить список заказов
      operationId: getOrders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/OrderPriority'
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersListResponse'

    post:
      tags: [Orders]
      summary: Создать заказ
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Получить заказ по ID
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Данные заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Orders]
      summary: Обновить заказ
      operationId: updateOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Заказ обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

    delete:
      tags: [Orders]
      summary: Удалить заказ
      operationId: deleteOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '204':
          description: Заказ удален
  /orders/{orderId}/status:
    patch:
      tags: [Orders]
      summary: Обновить статус заказа
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Статус обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OrderIdParam:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # --- Auth ---
...
    ProductStatus:
      type: string
      enum: [pending, processing, completed, rejected]
    Product:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: string
        name:
          type: string
        brand:
          type: string
        country:
          type: string
        category:
          type: string
        image:
          type: string
        weight:
          type: string
        dimensions:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        location:
          type: string
        packageType:
          type: string
        containerBarcode:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrderStatus:
      type: string
      enum: [pending, in_progress, completed, cancelled, shipped, delivered]
    OrderPriority:
      type: string
      enum: [low, medium, high, urgent]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        priority:
          $ref: '#/components/schemas/OrderPriority'
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProduct'
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/Address'
        assignedTo:
          type: string
        totalAmount:
          type: number
          format: float
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderProduct:
      type: object
      properties:
        id:
          type: integer
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
        picked:
          type: boolean
        location:
          type: string
    OrdersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateOrderRequest:
      type: object
      required: [products]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        products:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
                minimum: 1
        priority:
          $ref: '#/components/schemas/OrderPriority'
        shippingAddress:
          $ref: '#/components/schemas/CreateAddressRequest'
        notes:
          type: string
    UpdateOrderRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
        priority:
          $ref: '#/components/schemas/OrderPriority'
        assignedTo:
          type: string
        notes:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/UpdateAddressRequest'

    # --- Picking ---
...
    Address:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [pickup, delivery, warehouse, billing]
        country:
          type: string
          example: "Россия"
        region:
          type: string
          example: "Московская область"
        city:
          type: string
          example: "Москва"
        street:
          type: string
          example: "ул. Ленина"
        building:
          type: string
          example: "15"
        apartment:
          type: string
          example: "42"
        postalCode:
          type: string
          example: "123456"
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        fullAddress:
          type: string
          description: Полный адрес одной строкой
        isDefault:
          type: boolean
    CreateAddressRequest:
      type: object
      required: [city, street, building, postalCode]
      properties:
        type:
          type: string
          enum: [pickup, delivery, warehouse, billing]
        country:
          type: string
          default: "Россия"
        region:
          type: string
        city:
          type: string
        street:
          type: string
        building:
          type: string
        apartment:
          type: string
        postalCode:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        isDefault:
          type: boolean
    UpdateAddressRequest:
      type: object
      properties:
        country:
          type: string
        region:
          type: string
        city:
          type: string
        street:
          type: string
        building:
          type: string
        apartment:
          type: string
        postalCode:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        isDefault:
          type: boolean
...